cleanup, remove zerocoin, split hash.h from util.h
